@@ -66,6 +66,6 @@ func writeStarsChart(stars []stargazer, w io.Writer) error {
 	return graph.Render(chart.SVG, w)
 }
 
-func intValueFormatter(v interface{}) string {
+func intValueFormatter(v any) string {
 	return fmt.Sprintf("%.0f", v)
 }
